+++ test/triage/ok_then_retry_test.exs
@@ -1,3 +1,0 @@
-defmodule Triage.ThenTest do
-  use ExUnit.Case
-end